start teasing out the module structure
