+++ persistent/Database/Persist/Types/Base.hs
@@ -42,6 +42,9 @@ import Language.Haskell.TH.Syntax (Lift(..))
     -- instance on pre-1.2.4 versions of `text`
 import Instances.TH.Lift ()
 
+import Database.Persist.Types.FieldDef.Internal
+import Database.Persist.Types.Names
+
 -- | A 'Checkmark' should be used as a field type whenever a
 -- uniqueness constraint should guarantee that a certain kind of
 -- record may appear at most once, but other kinds of records may
@@ -119,29 +122,6 @@ data IsNullable = Nullable !WhyNullable
 data WhyNullable = ByMaybeAttr
                  | ByNullableAttr
                   deriving (Eq, Show)
-
--- | Convenience operations for working with '-NameDB' types.
---
--- @since 2.12.0.0
-class DatabaseName a where
-  escapeWith :: (Text -> str) -> (a -> str)
-
--- | An 'EntityNameDB' represents the datastore-side name that @persistent@
--- will use for an entity.
---
--- @since 2.12.0.0
-newtype EntityNameDB = EntityNameDB { unEntityNameDB :: Text }
-  deriving (Show, Eq, Read, Ord, Lift)
-
-instance DatabaseName EntityNameDB where
-  escapeWith f (EntityNameDB n) = f n
-
--- | An 'EntityNameHS' represents the Haskell-side name that @persistent@
--- will use for an entity.
---
--- @since 2.12.0.0
-newtype EntityNameHS = EntityNameHS { unEntityNameHS :: Text }
-  deriving (Show, Eq, Read, Ord, Lift)
 
 -- | An 'EntityDef' represents the information that @persistent@ knows
 -- about an Entity. It uses this information to generate the Haskell
@@ -266,68 +246,6 @@ data FieldType
     -- ^ Optional module and name.
     | FTApp FieldType FieldType
     | FTList FieldType
-    deriving (Show, Eq, Read, Ord, Lift)
-
--- | An 'EntityNameDB' represents the datastore-side name that @persistent@
--- will use for an entity.
---
--- @since 2.12.0.0
-newtype FieldNameDB = FieldNameDB { unFieldNameDB :: Text }
-    deriving (Show, Eq, Read, Ord, Lift)
-
--- | @since 2.12.0.0
-instance DatabaseName FieldNameDB where
-    escapeWith f (FieldNameDB n) = f n
-
--- | A 'FieldNameHS' represents the Haskell-side name that @persistent@
--- will use for a field.
---
--- @since 2.12.0.0
-newtype FieldNameHS = FieldNameHS { unFieldNameHS :: Text }
-    deriving (Show, Eq, Read, Ord, Lift)
-
--- | A 'FieldDef' represents the inormation that @persistent@ knows about
--- a field of a datatype. This includes information used to parse the field
--- out of the database and what the field corresponds to.
-data FieldDef = FieldDef
-    { fieldHaskell   :: !FieldNameHS
-    -- ^ The name of the field. Note that this does not corresponds to the
-    -- record labels generated for the particular entity - record labels
-    -- are generated with the type name prefixed to the field, so
-    -- a 'FieldDef' that contains a @'FieldNameHS' "name"@ for a type
-    -- @User@ will have a record field @userName@.
-    , fieldDB        :: !FieldNameDB
-    -- ^ The name of the field in the database. For SQL databases, this
-    -- corresponds to the column name.
-    , fieldType      :: !FieldType
-    -- ^ The type of the field in Haskell.
-    , fieldSqlType   :: !SqlType
-    -- ^ The type of the field in a SQL database.
-    , fieldAttrs     :: ![FieldAttr]
-    -- ^ User annotations for a field. These are provided with the @!@
-    -- operator.
-    , fieldStrict    :: !Bool
-    -- ^ If this is 'True', then the Haskell datatype will have a strict
-    -- record field. The default value for this is 'True'.
-    , fieldReference :: !ReferenceDef
-    , fieldCascade :: !FieldCascade
-    -- ^ Defines how operations on the field cascade on to the referenced
-    -- tables. This doesn't have any meaning if the 'fieldReference' is set
-    -- to 'NoReference' or 'SelfReference'. The cascade option here should
-    -- be the same as the one obtained in the 'fieldReference'.
-    --
-    -- @since 2.11.0
-    , fieldComments  :: !(Maybe Text)
-    -- ^ Optional comments for a 'Field'. There is not currently a way to
-    -- attach comments to a field in the quasiquoter.
-    --
-    -- @since 2.10.0
-    , fieldGenerated :: !(Maybe Text)
-    -- ^ Whether or not the field is a @GENERATED@ column, and additionally
-    -- the expression to use for generation.
-    --
-    -- @since 2.11.0.0
-    }
     deriving (Show, Eq, Read, Ord, Lift)
 
 isFieldNotGenerated :: FieldDef -> Bool
