+++ persistent/Database/Persist/Types/FieldDef.hs
@@ -1,0 +1,6 @@
+module Database.Persist.Types.FieldDef
+    (
+    ) where
+
+import Database.Persist.Types.FieldDef.Internal
+
